Code cleanup
- Remove unneeded editModel() method.
- Abstract out common code in factories that load files.
- Some JavaDoc improvements.
